[5] Move render to helpers
