+++ src/helpers/render.js
@@ -1,0 +1,14 @@
+const RenderPosition = {
+  BEFORE_END: 'beforeend',
+  AFTER_END: 'afterend',
+};
+
+const render = (
+  container,
+  template,
+  position = RenderPosition.BEFORE_END,
+) => {
+  container.insertAdjacentHTML(position, template);
+};
+
+export {render, RenderPosition};
