+++ src/main.js
@@ -10,11 +10,7 @@ import {getUser} from './mocks/user';
 import {getComments} from './mocks/comments';
 import {getDetailsModalTemplate} from './views/details-modal';
 import {getFilters} from './mocks/filters';
-
-const RenderPosition = {
-  BEFORE_END: 'beforeend',
-  AFTER_END: 'afterend',
-};
+import {render, RenderPosition} from './helpers/render';
 
 const MoviesCount = {
   ALL: 20,
@@ -29,14 +25,6 @@ const comments = getComments(COMMENTS_COUNT);
 const movies = getMovies(MoviesCount.ALL, comments);
 const user = getUser(movies);
 const filters = getFilters(movies, user);
-
-const render = (
-  container,
-  template,
-  position = RenderPosition.BEFORE_END,
-) => {
-  container.insertAdjacentHTML(position, template);
-};
 
 const headerElement = document.querySelector('.header');
